[PERF] mail: speed up fetching of `avatar_128`
Description:
Following d6c6351c7f0b241578121d459943fdf2b305571e, the `ir.rule`
related to discuss channels got more elaborated, with more subqueries.
When loading an `avatar_128`, in an override
of `_find_record_check_access`, a repetitive domain with an `any`
operator on the same left leaf. This repetitive domain + the new
`ir.rule` lead to a bad postgres plan, leading to a perf regression.

Fix:
We add an `auto_join=True` on `parent_channel_id` to avoid doing an
additional nested sub-query coming from the `ir.rule`. This change
is safe, as the comodel is the same model as the current one, and an
user that has access to a subchannel, should probably already have
access to its parent channel.
The other improvement is an optimistic early `super()` call, which
should not raise any issue for internal employees, who already
should have access to the avatar. In case of an access violation, we
fallback on the previous behaviour.

Benchmark:
On a large database, loading an `avatar_128`, cache disabled in
browser (simulating a first load) for an internal non-privileged user.

|                      | Timings   |
|----------------------|-----------|
| Before               | 300~400ms |
| `auto_join=True`     | 20~25ms   |
| + `super()` shortcut | 13~23ms   |

closes odoo/odoo#187498

X-original-commit: 12b80df4d7406656e4e49d1bb6271cce1fc09553
Signed-off-by: Piryns Victor (pivi) <pivi@odoo.com>