[FIX] web: list column widths: support width on button nodes
Before this commit, it was not possible to set the width attribute
in a list view arch on `<button>` nodes. This commit allows it.

One difficulty was that there's already some magic around button
columns. Indeed, adjacent buttons in the arch are gathered in a
single column. This doesn't map well with the width logic, as we
want to be able to define a width on a button, which will thus
define the width of its column. This can only work if the button
is alone in its column. For that reason, when a button has a width,
we do not group it with its potential adjacent buttons.

Task~4307553

closes odoo/odoo#187481

X-original-commit: 85ddc612a810a9c3a01eaa05c8e877e055928396
Signed-off-by: Francois Georis (fge) <fge@odoo.com>
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>